requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
